+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -238,12 +238,18 @@ public class ProjectManagerConst {
     public static final String EXPORTER_PARAM_QUERY_CONTEXT = "query-context";
     public static final String EXPORTER_PARAM_DEFAULT_OUTPUT_FORMAT = "query-default-output-format";
     public static final String EXPORTER_PARAM_DEFAULT_TEMPLATE_ID = "query-default-template-id";
+    public final static String EXPORTER_PARAM_QUERY_EXECUTION_ID = "query-execution-id";
     public static final String EXPORTER_QUERY_CONTEXT_PROJECT_ID = "PROJECT-ID";
+    public final static String EXPORTER_QUERY_CONTEXT_SEPARATOR = ";";
     public final static String API_KEY = "ApiKey";
+
 
     // Focus
     public final static String FOCUS_METADATA_PROJECT = "exporter";
-    public final static String FOCUS_TASK = "/v1/tasks";
+    public final static String FOCUS_TASK_PATH = "/v1/tasks";
+    public final static String FOCUS_TASK_RESULTS_PATH = "/results";
+    public final static String FOCUS_TASK_WAIT_TIME_PARAM = "wait_time";
+    public final static String FOCUS_TASK_WAIT_COUNT_PARAM = "wait_count";
 
     // Token Manager Variables
     public final static String TOKEN_MANAGER_ROOT = "/api";
@@ -297,6 +303,7 @@ public class ProjectManagerConst {
     public final static String TOKEN_MANAGER_URL = "TOKEN_MANAGER_URL";
     public final static String ENABLE_EMAILS = "ENABLE_EMAILS";
     public final static String MANAGE_TOKENS_CRON_EXPRESSION = "MANAGE_TOKENS_CRON_EXPRESSION";
+    public final static String EXPORTER_CRON_EXPRESSION = "EXPORTER_CRON_EXPRESSION";
     public final static String CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION = "CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION";
     public final static String EXPLORER_URL = "EXPLORER_URL";
     public final static String ENABLE_TOKEN_MANAGER = "ENABLE_TOKEN_MANAGER";
@@ -309,6 +316,8 @@ public class ProjectManagerConst {
     public final static String ENABLE_RSTUDIO_GROUP_MANAGER = "ENABLE_RSTUDIO_GROUP_MANAGER";
     public final static String OIDC_URL = "OIDC_URL";
     public final static String OIDC_REALM = "OIDC_REALM";
+    public final static String ENABLE_EXPORTER = "ENABLE_EXPORTER";
+    public final static String MAX_TIME_TO_WAIT_FOCUS_TASK_IN_MINUTES = "MAX_TIME_TO_WAIT_FOCUS_TASK_IN_MINUTES";
 
     // Spring Values (SV)
     public final static String HEAD_SV = "${";
@@ -337,6 +346,8 @@ public class ProjectManagerConst {
             HEAD_SV + WEBCLIENT_TCP_KEEP_INTERVAL_IN_SECONDS + ":60" + BOTTOM_SV;
     public final static String WEBCLIENT_TCP_KEEP_CONNECTION_NUMBER_OF_TRIES_SV =
             HEAD_SV + WEBCLIENT_TCP_KEEP_CONNECTION_NUMBER_OF_TRIES + ":10" + BOTTOM_SV;
+    public final static String MAX_TIME_TO_WAIT_FOCUS_TASK_IN_MINUTES_SV =
+            HEAD_SV + MAX_TIME_TO_WAIT_FOCUS_TASK_IN_MINUTES + ":5" + BOTTOM_SV;
     public final static String WEBCLIENT_MAX_NUMBER_OF_RETRIES_SV =
             HEAD_SV + WEBCLIENT_MAX_NUMBER_OF_RETRIES + ":3" + BOTTOM_SV;
     public final static String WEBCLIENT_TIME_IN_SECONDS_AFTER_RETRY_WITH_FAILURE_SV =
@@ -359,10 +370,13 @@ public class ProjectManagerConst {
     public final static String FOCUS_API_KEY_SV = HEAD_SV + FOCUS_API_KEY + BOTTOM_SV;
     public final static String ENABLE_EMAILS_SV = HEAD_SV + ENABLE_EMAILS + ":true" + BOTTOM_SV;
     public final static String ENABLE_TOKEN_MANAGER_SV = HEAD_SV + ENABLE_TOKEN_MANAGER + ":true" + BOTTOM_SV;
+    public final static String ENABLE_EXPORTER_SV = HEAD_SV + ENABLE_EXPORTER + ":true" + BOTTOM_SV;
     public final static String MANAGE_TOKENS_CRON_EXPRESSION_SV =
-            HEAD_SV + MANAGE_TOKENS_CRON_EXPRESSION + ":#{'0 0 * * * *'}" + BOTTOM_SV;
+            HEAD_SV + MANAGE_TOKENS_CRON_EXPRESSION + ":#{'0 * * * * *'}" + BOTTOM_SV;
+    public final static String EXPORTER_CRON_EXPRESSION_SV =
+            HEAD_SV + EXPORTER_CRON_EXPRESSION + ":#{'45 * * * * *'}" + BOTTOM_SV;
     public final static String CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION_SV =
-            HEAD_SV + CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION + ":#{'0 0 1,13 * * *'}" + BOTTOM_SV;
+            HEAD_SV + CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION + ":#{'30 * * * * *'}" + BOTTOM_SV;
     public final static String EXPLORER_URL_SV = HEAD_SV + EXPLORER_URL + BOTTOM_SV;
     public final static String EXPLORER_REDIRECT_URI_PARAMETER_SV = HEAD_SV + EXPLORER_REDIRECT_URI_PARAMETER + BOTTOM_SV;
     public final static String FRONTEND_PROJECT_CONFIG_SV = HEAD_SV + FRONTEND_PROJECT_CONFIG + BOTTOM_SV;
@@ -386,5 +400,6 @@ public class ProjectManagerConst {
     public final static String OIDC_EMAIL_CLAIM = "email";
     public final static String CUSTOM_PROJECT_CONFIGURATION = "CUSTOM";
     public final static String EMAIL_SERVICE = "EMAIL_SERVICE";
+
 
 }