+++ src/main/java/de/samply/exporter/focus/FocusServiceException.java
@@ -4,4 +4,9 @@ public class FocusServiceException extends Exception{
     public FocusServiceException(String message) {
         super(message);
     }
+
+    public FocusServiceException(Throwable cause) {
+        super(cause);
+    }
+
 }