+++ src/main/java/de/samply/frontend/dto/ProjectBridgehead.java
@@ -1,6 +1,7 @@
 package de.samply.frontend.dto;
 
 import de.samply.project.state.ProjectBridgeheadState;
+import de.samply.query.QueryState;
 
 import java.time.Instant;
 
@@ -8,6 +9,7 @@ public record ProjectBridgehead(
         String projectCode,
         String bridgehead,
         ProjectBridgeheadState state,
-        Instant modifiedAt
+        Instant modifiedAt,
+        QueryState queryState
 ) {
 }