+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-04-15]
+## [0.0.1 - 2024-04-18]
 ### Added
 - First version of the project
 - Spring Application
@@ -130,3 +130,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Create notification after sending an email
 - Send message while requesting changes in script or project
 - Send message while rejecting a project
+- Exporter Job
+- Check Queries in Exporter Job
+- Max time to wait focus task in minutes